Introduce approach to split TypeId and TypeDef data structures
This introduces a new concept where TypeId and TypeDef data structures
do now have a so-called "FreeForm" and "CompactForm".
The free form is free of dependencies and can and should be constructed by HasTypeId and HasTypeDef
whereas the compact form is stored within the type registry for a more compact result upon serialization.

This approach enables us to use a single data structure for both purposes.